refactor: useValidate(지금은 사용x)도 타입스크립트로 바꿔보기
